Add string testers to component
These test a string according to a certain criteria and
return either true or false.

The goal is to have some commonly useful testers in
this component. So far we just have the valid UTF8
and the valid datetime testers, but additional testers
for things like IPs, domains, emails etc. could be
helpful too. Some of the string filters could be adapted
to have a tester version, for when you just want to find
out if a string is valid not forceably change it.