Better support typed properties
Also removes compatibility to diverging property paths defined
in the form, we only support direct properties which are
mapped 1:1 in the form. This seems easiest and anything else
would just lead to a headache, although maybe there is a better
option sometime in the future or when we go PHP 8 only.

The PHP 8 support is preliminary, as it was not tested yet.